+++ tests/testthat/test-tabraux.R
@@ -1,34 +1,4 @@
-## tests for pc2
-i1 = c()
-i2 = ""
-i3 = "a b c d e f g"
-i4 = "a b c d e f g x"
-i5 = c("","")
-i6 = c("",i3)
-i7 = c(i3,i4)
-i8 = " "
-i9 = c(1953, 2019)
-test_that("pc2 identical to tabr::pc when no redundant blanks", {
-  expect_identical(pc2(i1),tabr::pc(i1) )
-  expect_identical(pc2(i2),tabr::pc(i2) )
-  expect_identical(pc2(i3),tabr::pc(i3) )
-  expect_identical(pc2(i4),tabr::pc(i4) )
-  expect_identical(pc2(i5),tabr::pc(i5) )
-  expect_identical(pc2(i6),tabr::pc(i6) )
-  expect_identical(pc2(i7),tabr::pc(i7) )
-  expect_identical(pc2(i8),tabr::pc(i8) )
-  expect_identical(pc2(i9),tabr::pc(i9) )
-})
-i3a = "  a b c d e f g"
-i3b = "a b   c d  e f  g"
-i3c = "a b   c d  e f  g  "
-i8a = "   "
-test_that("pc2 identical to tabr::pc when redundant blanks are removed", {
-  expect_identical(pc2(i3a),tabr::pc(i3) )
-  expect_identical(pc2(i3b),tabr::pc(i3) )
-  expect_identical(pc2(i3c),tabr::pc(i3) )
-  expect_identical(pc2(i8a),tabr::pc(i8) )
-})
+
 ## tests for expand_notes
 i9 = 'a3  b r c4 g3 f'
 o9 = "a3 b3 r c#4 g3 f#3"
@@ -64,4 +34,4 @@ test_that("check_times", {
   expect_identical(o16$times,"2. 8*3 16*4 8 8*6 8*3 16*4 8" )
   expect_identical(o16$counts,rep(6,4) )
 })
-
+## tests for lilypond_version